Feat: 주차장 전체리스트 가격순 정렬 완료

+++ AOS/app/src/main/java/com/team/parking/data/model/map/MapOrderResponse.kt
@@ -7,9 +7,9 @@ data class MapOrderResponse(
     @SerializedName("feeBasic")
     val feeBasic: Int,
     @SerializedName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerializedName("lng")
-    val lng: Int,
+    val lng: Double,
     @SerializedName("meter")
     val meter: Int,
     @SerializedName("parkId")
